moved the add console dialog to the dialog view folder
