+++ app/app/src/main/java/com/emilygranville/videogamelist/View/Dialogs/AddConsoleDialog.java
@@ -1,4 +1,4 @@
-package com.emilygranville.videogamelist.View;
+package com.emilygranville.videogamelist.View.Dialogs;
 
 import android.app.Dialog;
 import android.os.Bundle;
